# api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Neoconnect",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Dev version with docker-compose and nginx proxy"
    },
    {
      "url": "/web",
      "description": "Production version"
    }
  ],
...
  "paths": {
...
    "/shop/search": {
      "post": {
        "tags": [
          "Shop"
        ],
        "summary": "Search for Shop related to a keyword",
        "parameters": [
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pseudo": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "pseudo": "jordan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request Body Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "Error Parameters": {
                    "value": {
                      "code": 400,
                      "message": "Mauvais paramètre"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/shop/me": {
      "get": {
        "tags": [
          "Shop"
        ],
        "summary": "Get Information of current account",
        "parameters": [
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "Shop"
        ],
        "summary": "Edit Information of current account",
        "parameters": [
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "description": "Request Body Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "Error Body": {
                    "value": {
                      "code": 400,
                      "message": "Erreur dans un champ"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/shop/{id}": {
      "get": {
        "tags": [
          "Shop"
        ],
        "summary": "Get Information of another account",
        "parameters": [
          {
            "$ref": "#/components/parameters/idUser"
          },
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request Body Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "Id non existant": {
                    "value": {
                      "code": 400,
                      "message": "Id non existant"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/shop/listInf": {
      "get": {
        "tags": [
          "Shop"
        ],
        "summary": "Get Information on all influencers",
        "parameters": [
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InfResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/shop/follow/{id}": {
      "put": {
        "tags": [
          "Follow"
        ],
        "summary": "Permet de follow un shop et de recevoir des mails a chaque nouvelle offre cree",
        "parameters": [
          {
            "$ref": "#/components/parameters/idUser"
          },
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "responses": {
          "200": {
            "description": "Follow Success"
          },
          "400": {
            "description": "Request Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "Error Body": {
                    "value": {
                      "code": 400,
                      "message": "Erreur dans un champ"
                    }
                  },
                  "Id non existant": {
                    "value": {
                      "code": 400,
                      "message": "Id non existant"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Follow"
        ],
        "summary": "Permet de recuperer la liste des followers du shop",
        "parameters": [
          {
            "$ref": "#/components/parameters/idUser"
          },
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "responses": {
          "200": {
            "description": "liste des followers du Shop",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "idUser": {
                      "type": "number"
                    },
                    "idFollow": {
                      "type": "number"
                    },
                    "id": {
                      "type": "number"
                    },
                    "pseudo": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request Body Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "Id non existant": {
                    "value": {
                      "code": 400,
                      "message": "Id non existant"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/shop/unfollow/{id}": {
      "put": {
        "tags": [
          "Follow"
        ],
        "summary": "Permet de unfollow un shop et de ne plus recevoir des mails a chaque nouvelle offre cree",
        "parameters": [
          {
            "$ref": "#/components/parameters/idUser"
          },
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "responses": {
          "200": {
            "description": "Unfollow Success"
          },
          "400": {
            "description": "Request Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "Error Body": {
                    "value": {
                      "code": 400,
                      "message": "Erreur dans un champ"
                    }
                  },
                  "Id non existant": {
                    "value": {
                      "code": 400,
                      "message": "Id non existant"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/user/follow": {
      "get": {
        "tags": [
          "Follow"
        ],
        "summary": "Permet de recuperer la liste des shop follow par l'influencer",
        "parameters": [
          {
            "$ref": "#/components/parameters/Header"
          }
        ],
        "responses": {
          "200": {
            "description": "liste des follow de l'influencer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "idUser": {
                      "type": "number"
                    },
                    "idFollow": {
                      "type": "number"
                    },
                    "id": {
                      "type": "number"
                    },
                    "pseudo": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "Id non existant": {
                    "value": {
                      "code": 400,
                      "message": "Id non existant"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "204": {
        "description": "Success"
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": 401,
              "message": "User non reconnu / Mauvais Token"
            }
          }
        }
      },
...
    },
    "parameters": {
      "idUser": {
        "name": "id",
        "in": "path",
        "description": "user ID",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/ID"
        }
      },
...
      "Header": {
        "name": "authorization",
        "in": "header",
        "description": "Bearer Token",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Header"
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Header": {
        "type": "object",
        "properties": {
          "authorization": {
            "type": "string"
          }
        }
      },
      "ID": {
        "type": "string"
      },
...
      "InfResponse": {
        "type": "object",
        "properties": {
          "pseudo": {
            "type": "string"
          },
          "sexe": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "userPicture": {
            "type": "string"
          },
          "theme": {
            "type": "string",
            "enum": [
              "mode",
              "high tech",
              "jeux videos",
              "cosmetique",
              "food",
              "sport/fitness"
            ]
          },
          "facebook": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "snapchat": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "youtube": {
            "type": "string"
          },
          "mark": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "comment": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ShopRequest": {
        "type": "object",
        "properties": {
          "pseudo": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "userPicture": {
            "type": "string"
          },
          "theme": {
            "type": "string",
            "enum": [
              "mode",
              "high tech",
              "jeux videos",
              "cosmetique",
              "food",
              "sport/fitness"
            ]
          },
          "society": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "snapchat": {
            "type": "string"
          }
        },
        "required": [
          "pseudo",
          "password"
        ]
      },
      "ShopResponse": {
        "type": "object",
        "properties": {
          "pseudo": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "userPicture": {
            "type": "string"
          },
          "theme": {
            "type": "string",
            "enum": [
              "mode",
              "high tech",
              "jeux videos",
              "cosmetique",
              "food",
              "sport/fitness"
            ]
          },
          "society": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "snapchat": {
            "type": "string"
          },
          "mark": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "comment": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SearchUserResponse": {
        "type": "object",
        "properties": {
          "userType": {
            "type": "string"
          },
          "pseudo": {
            "type": "string"
          },
          "userPicture": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "theme": {
            "type": "string"
          },
          "note": {
            "type": "number"
          }
        }
      },
...
    }
  }
}